create unit tests for CommitCommentReactionClient
